:art: Refactor, make everything base classes, drop treatment
